refactor(g4f/Provider/CablyAI.py > g4f/Provider/needs_auth/CablyAI.py): restructure provider and simplify model handling
