+++ g4f/Provider/needs_auth/__init__.py
@@ -1,5 +1,6 @@
 from .Anthropic         import Anthropic
 from .BingCreateImages  import BingCreateImages
+from .CablyAI           import CablyAI
 from .Cerebras          import Cerebras
 from .CopilotAccount    import CopilotAccount
 from .Custom            import Custom
